function getUsers.js
