+++ burger-queen/src/Components/Admin/ListUsers.js
@@ -1,14 +1,24 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { useState } from "react";
+import { getUsers } from "../../util/getUsers";
 
 
 export const ListUsers = () => {
 
-const [createUser, setCreateUser] = useState({});
+const [createUser, setCreateUser] = useState([]);
+
+useEffect(() =>{
+    const showUserList = () => {
+        getUsers()
+          .then((res) => setCreateUser (res))
+          .catch((err) => console.log(err))
+    }
+    showUserList();
+},[])
 
 return (
 <section id='sectionTableUsers'>
-   <table className="table table-warning table-striped table-hover">
+   <table className="table table-warning table-striped table-hover showList">
     <thead>
         <tr>
             <th scope="col">UserID</th>
@@ -21,15 +31,20 @@ return (
         </tr>
     </thead>
     <tbody>
-        <tr>
-            <th scope="row">1</th>
-            <td>Mark</td>
-            <td>mesero</td>
-            <td>@mdo</td>
-            <td>*******</td>
-            <td><i className="fa-solid fa-pen-to-square"></i></td>
-            <td><i className="fa-regular fa-trash-can"></i></td>
-        </tr>
+      {createUser.map((user) => {
+          return (
+            <tr key={user.id}>
+                <th scope="row">{user.id}</th>
+                <td>{user.fullName}</td>
+                <td>{user.userRol}</td>
+                <td>{user.email}</td>
+                <td>{user.password}</td>
+                <td><i className="fa-solid fa-pen-to-square"></i></td>
+                <td><i className="fa-regular fa-trash-can"></i></td>
+            </tr>
+            
+            )}
+        )}
     </tbody>
    </table>
 </section>
